cli-client/socket/requirements.txt
python-socketio[asyncio_client]
pyfiglet
requests
textual